requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
click==7.1.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
pathspec==0.7.0
pickleDB==0.9.2
pylint==2.4.4
pyzmq==19.0.0
regex==2020.4.4
six==1.14.0
tinydb==3.15.2
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.11.2
